+++ app/components/Sidebar/utils.tsx
@@ -1,13 +1,18 @@
 import { renameFile } from "@tauri-apps/api/fs";
 import {
   SiC,
+  SiCss3,
   SiGo,
   SiHtml5,
-  SiImagej,
+  SiImgur,
   SiJavascript,
+  SiJson,
+  SiMarkdown,
   SiPython,
   SiReact,
   SiRust,
+  SiSass,
+  SiSvg,
   SiTypescript,
 } from "react-icons/si";
 import { TiDocument } from "react-icons/ti";
@@ -22,23 +27,37 @@ export const Rename = async (path: string, Name: string) => {
 export const LanguageLogo = ({ suffix }: { suffix: string }) => {
   switch (suffix) {
     case "html":
-      return <SiHtml5 />;
+      return <SiHtml5 className="text-red-500" />;
+    case "css":
+      return <SiCss3 className="text-blue-400" />;
+    case "scss":
+      return <SiSass className="text-pink-600" />;
     case "js":
-      return <SiJavascript />;
+      return <SiJavascript className="text-yellow-400" />;
+    case "mjs":
+      return <SiJavascript className="text-yellow-400" />;
     case "ts":
-      return <SiTypescript />;
+      return <SiTypescript className="text-blue-400" />;
+    case "jsx":
+      return <SiReact className="text-blue-500" />;
     case "tsx":
-      return <SiReact />;
+      return <SiReact className="text-blue-500" />;
     case "rs":
-      return <SiRust />;
+      return <SiRust className="text-orange-500" />;
     case "go":
-      return <SiGo />;
+      return <SiGo className="text-blue-500" />;
     case "py":
-      return <SiPython />;
+      return <SiPython className="text-yello-400" />;
     case "c":
-      return <SiC />;
+      return <SiC className="text-blue-500" />;
+    case "md":
+      return <SiMarkdown className="text-blue-500" />;
+    case "json":
+      return <SiJson className="text-green-600" />;
     case "png":
-      return <SiImagej />;
+      return <SiImgur className="text-green-700" />;
+    case "svg":
+      return <SiSvg className="text-yello-500" />;
     default:
       return <TiDocument />;
   }