+++ app/components/index.ts
@@ -3,5 +3,7 @@ import * as Sidebar from "./Sidebar";
 import * as Collapse from "./ui/collapsible";
 import * as Context from "./ui/context-menu";
 import * as Resize from "./ui/resizable";
+import * as Tab from "./ui/tabs";
+import Menu from "./Menu";
 
-export { Editor, Sidebar, Collapse, Context, Resize };
+export { Editor, Sidebar, Collapse, Context, Resize, Tab, Menu };